requirements.txt
Django==3.2.9
psycopg2-binary==2.9.1
pytz==2021.3
black==24.2.0
isort==5.13.2
pylint==3.0.3